// types/user.ts
export default interface User {
    userName: string;
    imageFilenames: any[];
    tier: string;
    occupation: string;
    state: string;
    city: string;
    bio: string;
    matchSex: string;
    apiKey: string;
    birthDate: string;
    createdAt: string;
    email: string;
    firstName: string;
    privateProfile: boolean;
    sex: string;
    firstLogin: boolean;
}